[skip ci] Revert "Add Action to add cc'ed people as reviewers (#9934)"
This reverts commit bae144cfa51b5ff176e45e337d15d299bc8133b2.
