--- .github/workflows/cc_bot.yml
@@ -1,46 +1,0 @@
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-
-# GH actions.
-# We use it to cover windows and mac builds
-# Jenkins is still the primary CI
-
-name: PR
-
-on:
-  # See https://docs.github.com/en/actions/using-workflows/events-that-trigger-workflows#pull_request_target
-  pull_request_target:
-    types: [assigned, opened, synchronize, reopened, edited, ready_for_review]
-
-concurrency:
-  group: PR-${{ github.event.pull_request.number }}
-  cancel-in-progress: true
-
-jobs:
-  cc-reviewers:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-        with:
-          submodules: "recursive"
-      - name: Add cc'ed reviewers
-        env:
-          PR: ${{ toJson(github.event.pull_request) }}
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-        run: |
-          set -eux
-          python tests/scripts/github_cc_reviewers.py
